Stop setting CONFIG_MAXSMP and adjust NR_CPUS
During the Kconfig factor, CONFIG_MAXSMP was set for all builds instead
of just debug builds. This bumps the NR_CPUs up to 8192 which is far
more than anyone actually needs for Fedora. Stop setting CONFIG_MAXSMP
on non-debug configs and adjust NR_CPUs back to what it was before.